py/requirements.txt
pandas==2.2.2
openpyxl==3.1.5
azure-storage-blob==12.19.1
python-dotenv==1.0.1
